+++ test/Chord1LTest.java
@@ -1,0 +1,14 @@
+/**
+ * Customized JUnit test fixture for {@code Chord1L}.
+ */
+public class Chord1LTest extends ChordTest {
+
+    @Override
+    protected final Chord constructorTest() {
+
+        Chord test = new Chord1L();
+        return test;
+
+    }
+
+}